Rename robotoff.ml into robotoff.prediction
