+++ robotoff/prediction/category/prediction_from_ocr/constants.py
