+++ robotoff/prediction/object_detection/core.py
@@ -7,9 +7,9 @@ import PIL
 from PIL import Image
 
 from robotoff import settings
-from robotoff.ml.object_detection.utils import label_map_util
-from robotoff.ml.object_detection.utils import visualization_utils as vis_util
-from robotoff.ml.object_detection.utils.label_map_util import CategoryIndex
+from robotoff.prediction.object_detection.utils import label_map_util
+from robotoff.prediction.object_detection.utils import visualization_utils as vis_util
+from robotoff.prediction.object_detection.utils.label_map_util import CategoryIndex
 from robotoff.utils import get_logger, http_session
 from robotoff.utils.types import JSONType
