+++ robotoff/prediction/object_detection/download.py
