+++ robotoff/prediction/object_detection/utils/label_map_util.py
