+++ robotoff/prediction/object_detection/utils/string_int_label_map.proto
