+++ robotoff/prediction/object_detection/utils/visualization_utils.py
