+++ tests/prediction/category/neural/test_category_classifier.py
@@ -4,7 +4,7 @@ import pytest
 
 from robotoff.insights._enum import InsightType
 from robotoff.insights.dataclass import RawInsight
-from robotoff.ml.category.neural.category_classifier import (
+from robotoff.prediction.category.neural.category_classifier import (
     CategoryClassifier,
     Prediction,
 )
@@ -45,9 +45,7 @@ def _prediction_resp(categories: List[str], confs: List[float]) -> MockResponse:
 def test_predict_missing_data():
     classifier = CategoryClassifier(None)
 
-    predicted = classifier.predict(
-        {"WRONG_ingredients_tags": ["ingredient1"]},
-    )
+    predicted = classifier.predict({"WRONG_ingredients_tags": ["ingredient1"]},)
 
     assert not predicted
 
@@ -76,20 +74,10 @@ def test_predict_missing_data():
 def test_predict(mocker, deepest_only, mock_response, want_predictions):
     category_taxonomy = Taxonomy.from_dict(
         {
-            "en:meat": {
-                "names": "meat",
-            },
-            "en:fish": {
-                "names": "fish",
-            },
-            "en:salmon": {
-                "names": "salmon",
-                "parents": ["en:fish"],
-            },
-            "en:smoked-salmon": {
-                "names": "salmon",
-                "parents": ["en:salmon"],
-            },
+            "en:meat": {"names": "meat",},
+            "en:fish": {"names": "fish",},
+            "en:salmon": {"names": "salmon", "parents": ["en:fish"],},
+            "en:smoked-salmon": {"names": "salmon", "parents": ["en:salmon"],},
         }
     )
