+++ tests/workers/tasks/test_product_updated.py
@@ -1,7 +1,7 @@
 from robotoff import settings
 from robotoff.insights._enum import InsightType
 from robotoff.insights.dataclass import ProductInsights, RawInsight
-from robotoff.ml.category.neural.category_classifier import Prediction
+from robotoff.prediction.category.neural.category_classifier import Prediction
 from robotoff.workers.tasks.product_updated import add_category_insight
 
 # TODO: refactor function under test to make it easier to test
